+++ rolling/__init__.py
@@ -1,6 +1,7 @@
 from .apply import Apply
 from .arithmetic import Sum, Product, Nunique
 from .entropy import Entropy
+from .hash import PolynomialHash
 from .logical import All, Any
 from .minmax import Min, Max, MinHeap
 from .stats import Mean, Var, Std, Median, Mode, Skew, Kurtosis